fix: change the commandDir from quorum to eth, and rename test/quorum to test/eth
