+++ src/eth/command/explorer.ts
@@ -2,7 +2,7 @@ import { Argv } from 'yargs'
 
 export const command = 'explorer'
 
-export const desc = '管理 Quorum explorer 的指令'
+export const desc = '管理 Eth explorer 的指令'
 
 export const builder = (yargs: Argv) => {
   return yargs.commandDir('explorer').demandCommand()
