+++ src/eth/command/network.ts
@@ -2,7 +2,7 @@ import { Argv } from 'yargs'
 
 export const command = 'network'
 
-export const desc = '管理 Quorum network 的指令'
+export const desc = '管理 Eth network 的指令'
 
 export const builder = (yargs: Argv) => {
   return yargs.commandDir('network').demandCommand()
