+++ test/eth/service/network.test.ts
@@ -2,12 +2,12 @@
 import Dockerode from 'dockerode'
 import assert from 'assert'
 import sinon from 'sinon'
-import Network from '../../../src/quorum/service/network'
-import config from '../../../src/quorum/config'
+import Network from '../../../src/eth/service/network'
+import config from '../../../src/eth/config'
 import fs from 'fs'
 import { resolve } from 'path'
 import { sleep, TimeLimitError } from '../../../src/util'
-import { JoinNodeType } from '../../../src/quorum/model/type/network.type'
+import { JoinNodeType } from '../../../src/eth/model/type/network.type'
 import Wallet from '../../../src/wallet/service/wallet'
 import { WalletType } from '../../../src/wallet/model/type/wallet.type'
